+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [4.3.X.X] - 2025-XX-XX
+### Changed
+- Show point in the icon of opsi module validation to indicate a warning or an expired license instead of changing color of the icon
+
+### Add
+- Show a point in the Health Check icon to show if there is a warning/error in the Health Check
+
 ## [4.3.8.1] - 2025-03-27
 ### Fixed
 - Icon to select all products appears when depot selection changes (even if all products are selected)